+++ src/main/java/org/opensearch/security/action/configupdate/ConfigUpdateRequest.java
@@ -28,6 +28,7 @@ package org.opensearch.security.action.configupdate;
 
 import java.io.IOException;
 
+import org.opensearch.Version;
 import org.opensearch.action.ActionRequestValidationException;
 import org.opensearch.action.support.nodes.BaseNodesRequest;
 import org.opensearch.core.common.io.stream.StreamInput;
@@ -36,10 +37,14 @@ import org.opensearch.core.common.io.stream.StreamOutput;
 public class ConfigUpdateRequest extends BaseNodesRequest<ConfigUpdateRequest> {
 
     private String[] configTypes;
+    private String[] entityNames;
 
     public ConfigUpdateRequest(StreamInput in) throws IOException {
         super(in);
         this.configTypes = in.readStringArray();
+        if (in.getVersion().onOrAfter(Version.V_3_1_0)) {
+            this.entityNames = in.readOptionalStringArray();
+        }
     }
 
     public ConfigUpdateRequest() {
@@ -51,10 +56,19 @@ public class ConfigUpdateRequest extends BaseNodesRequest<ConfigUpdateRequest> {
         setConfigTypes(configTypes);
     }
 
+    public ConfigUpdateRequest(String configType, String[] entityNames) {
+        this();
+        setConfigTypes(new String[] { configType });
+        setEntityNames(entityNames);
+    }
+
     @Override
     public void writeTo(final StreamOutput out) throws IOException {
         super.writeTo(out);
         out.writeStringArray(configTypes);
+        if (out.getVersion().onOrAfter(Version.V_3_1_0)) {
+            out.writeOptionalStringArray(entityNames);
+        }
     }
 
     public String[] getConfigTypes() {
@@ -65,9 +79,19 @@ public class ConfigUpdateRequest extends BaseNodesRequest<ConfigUpdateRequest> {
         this.configTypes = configTypes;
     }
 
+    public String[] getEntityNames() {
+        return entityNames;
+    }
+
+    public void setEntityNames(final String[] entityNames) {
+        this.entityNames = entityNames;
+    }
+
     @Override
     public ActionRequestValidationException validate() {
         if (configTypes == null || configTypes.length == 0) {
+            return new ActionRequestValidationException();
+        } else if (configTypes.length > 1 && (entityNames != null && entityNames.length > 1)) {
             return new ActionRequestValidationException();
         }
         return null;